refactor(store): flatten upcasting/ into upcasting.rs
Merge upcasting/morsel.rs (EventMorsel) and upcasting/upcaster.rs (Upcaster
trait + no-op () impl) into a single upcasting.rs at the crate root. The
two types form one cohesive concept — the schema-migration pipeline — and
splitting them across files served no purpose now that the directory has
no other contents.

Public paths nexus_store::upcasting::EventMorsel and ::Upcaster unchanged.
