--- crates/nexus-store/src/upcasting/mod.rs
@@ -1,5 +1,0 @@
-mod morsel;
-mod upcaster;
-
-pub use morsel::EventMorsel;
-pub use upcaster::Upcaster;
